% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fooRanges.R
\name{barbar}
\alias{barbar}
\alias{barbar,fooRanges-method}
\title{Accessor for barbar of fooRanges}
\usage{
barbar(x, ...)

\S4method{barbar}{fooRanges}(x)
}
\arguments{
\item{x}{fooRanges object}

\item{...}{additional arguments}
}
\value{
the barbar
}
\description{
A function that gets the barbar from fooRanges
}
