% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fooRanges.R
\name{foobar}
\alias{foobar}
\alias{foobar,fooRanges-method}
\title{Foobar fooRanges}
\usage{
foobar(x, ...)

\S4method{foobar}{fooRanges}(x)
}
\arguments{
\item{x}{fooRanges object}

\item{...}{additional arguments}
}
\value{
a summary
}
\description{
A function that does foobar to fooRanges
}
\examples{

library(GenomicRanges)
x <- GRanges("1", IRanges(1,10))
y <- fooRanges(x, barbar="hi")
barbar(y)
barbar(y) <- "hey"
foobar(y)

}
