% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{fooRanges-class}
\alias{fooRanges-class}
\alias{fooRanges}
\title{fooRanges}
\usage{
fooRanges(x, barbar)
}
\arguments{
\item{x}{a GRanges object}

\item{barbar}{character string}
}
\value{
a fooRanges object
}
\description{
The fooRanges object and constructor
}
