% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-FooRanges.R
\name{barbar}
\alias{barbar}
\alias{barbar<-}
\alias{barbar,FooRanges-method}
\alias{barbar<-,FooRanges,character-method}
\title{Accessor and replacement methods for barbar in FooRanges}
\usage{
barbar(x, ...)

barbar(x, ...) <- value

\S4method{barbar}{FooRanges}(x)

\S4method{barbar}{FooRanges,character}(x) <- value
}
\arguments{
\item{x}{FooRanges object}

\item{...}{additional arguments}

\item{value}{replacement for barbar}
}
\description{
Functions that get or set the barbar slot from FooRanges
}
