% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-FooRanges.R
\name{foobar}
\alias{foobar}
\alias{foobar,FooRanges-method}
\title{foobar generic and methods}
\usage{
foobar(x, ...)

\S4method{foobar}{FooRanges}(x)
}
\arguments{
\item{x}{FooRanges object}

\item{...}{additional arguments}
}
\value{
a summary
}
\description{
A function that does foobar to FooRanges
}
\examples{

library(GenomicRanges)
x <- GRanges("1", IRanges(1,10))
y <- FooRanges(x, barbar="hi")
barbar(y)
barbar(y) <- "hey"
foobar(y)

}
