% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FooRanges-class}
\alias{FooRanges-class}
\alias{.FooRanges}
\alias{FooRanges}
\title{FooRanges}
\usage{
FooRanges(x, barbar)
}
\arguments{
\item{x}{a GRanges object}

\item{barbar}{character string}
}
\value{
a FooRanges object
}
\description{
The FooRanges object and constructor
}
