% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{FooRanges}
\alias{FooRanges}
\title{FooRanges}
\usage{
FooRanges(x, barbar)
}
\arguments{
\item{x}{a GRanges object}

\item{barbar}{character string}
}
\value{
a FooRanges class object
}
\description{
The FooRanges constructor function
}
\seealso{
\code{\link{FooRanges-class}}
}
